Added processing of ghost cells for the grid, as well as reconstruction of primitive states at cell interfaces

+++ src/fluid/cell.hpp
@@ -12,7 +12,7 @@
 // struct for each cell
 struct cell {
     // index
-    double i, j, k; // spatial index within patch
+    int i, j, k; // spatial index within patch
     // variables
     cons U;
     prim W;
